requirements.txt
streamlit==1.28.0
transformers==4.35.0
torch==2.1.0
opencv-python-headless==4.8.1.78
Pillow==10.0.1
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
plotly==5.17.0 